Added minor changes to have the game run on 'old item' instead of the new one to transition the new item refactor
